Split config defaults and validation
Closes #143
